+++ ui/src/colors.rs
@@ -1,29 +1,77 @@
-/// Color of the gridlines. This will be configurable in the future.
-pub const GRIDLINES: [f32; 4] = [0.25, 0.25, 0.25, 1.0];
+//! Color constants. All of these will be configurable in the future.
 
-/// Color given to the hovered cell when drawing. This will be configurable in
-/// the future.
-pub const HOVERED_DRAW: [f32; 4] = [0.0, 0.5, 1.0, 1.0];
+use palette::{Srgb, Srgba};
 
-/// Color given to the hovered cell when selecting. This will be configurable in
-/// the future.
-pub const HOVERED_SELECT: [f32; 4] = [1.0, 1.0, 1.0, 1.0];
+/// Defines an `Rgb` value at compile time.
+macro_rules! rgb {
+    ($r:expr, $g:expr, $b:expr) => {
+        palette::rgb::Rgb {
+            red: $r,
+            green: $g,
+            blue: $b,
+            standard: std::marker::PhantomData,
+        }
+    };
+}
+/// Defines an `Rgba` value at compile time.
+macro_rules! rgba {
+    ($r:expr, $g:expr, $b:expr, $a:expr) => {
+        palette::Alpha {
+            color: rgb!($r, $g, $b),
+            alpha: $a,
+        }
+    };
+}
 
-/// Color given to the selection resize preview. This will be configurable in
-/// the future.
-pub const SELECTION_RESIZE: [f32; 4] = [0.0, 0.5, 1.0, 0.4];
+/// 2D background color.
+pub const BACKGROUND_2D: Srgb = rgb!(0.0, 0.0, 0.0);
+/// 3D background color.
+pub const BACKGROUND_3D: Srgb = rgb!(0.0, 0.0, 0.0);
 
-/// Color given to the selection. This will be configurable in the future.
-pub const SELECTION: [f32; 4] = [0.6, 0.6, 0.6, 0.8];
+/// Color of the gridlines.
+pub const GRIDLINES: Srgba = rgba!(0.25, 0.25, 0.25, 1.0);
 
-/// 2D cell background color. This will be configurable in the future.
-pub const BACKGROUND_2D: [f32; 4] = [0.0, 0.0, 0.0, 1.0];
+/// Opacity of 2D crosshairs.
+pub const CROSSHAIR_OPACITY_2D: f32 = 0.75;
 
-/// 3D cell background color. This will be configurable in the future.
-pub const BACKGROUND_3D: [f32; 4] = [0.0, 0.0, 0.0, 1.0];
+pub mod hover {
+    use super::*;
 
-/// Color for dead cells. This will be configurable in the future.
-pub const DEAD: [u8; 4] = [0, 0, 0, 0];
+    /// Color of the hovered cell when drawing.
+    pub const DRAW_FILL: Srgba = rgba!(0.0, 0.25, 0.5, 0.75);
+    /// Color of the outline around the hovered cell when drawing.
+    pub const DRAW_OUTLINE: Srgb = rgb!(0.0, 0.5, 1.0);
 
-/// Color for live cells. This will be configurable in the future.
-pub const LIVE: [u8; 4] = [255, 255, 255, 255];
+    /// Color of the hovered cell when selecting.
+    pub const SELECT_FILL: Srgba = rgba!(0.6, 0.6, 0.6, 0.75);
+    /// Color of the outline around the hovered cell when selecting.
+    pub const SELECT_OUTLINE: Srgb = rgb!(0.8, 0.8, 0.8);
+}
+
+pub mod selection {
+    use super::*;
+
+    /// Color given to the region selection.
+    pub const REGION_FILL: Srgba = rgba!(0.6, 0.6, 0.6, 0.25);
+    /// Color of the outline around the region selection.
+    pub const REGION_OUTLINE: Srgb = rgb!(0.6, 0.6, 0.6);
+
+    /// Color of the cell selection.
+    pub const CELLS_FILL: Srgba = rgba!(0.6, 0.6, 0.6, 0.0);
+    /// Color of the outline around the cell selection.
+    pub const CELLS_OUTLINE: Srgb = rgb!(0.6, 0.6, 0.6);
+
+    /// Color of the selection resize preview.
+    pub const RESIZE_FILL: Srgba = rgba!(0.8, 0.4, 0.0, 0.125);
+    /// Color of the outline around the selection resize preview.
+    pub const RESIZE_OUTLINE: Srgb = rgb!(0.8, 0.4, 0.0);
+}
+
+pub mod cells {
+    use super::*;
+
+    /// Color for dead cells.
+    pub const DEAD: Srgba = rgba!(0.0, 0.0, 0.0, 0.0);
+    /// Color for live cells.
+    pub const LIVE: Srgba = rgba!(1.0, 1.0, 1.0, 1.0);
+}