+++ ui/src/gridview/render/render3d/fog.rs
@@ -1,10 +1,13 @@
 use super::*;
 
+/// Fog rendering parameters.
+///
+/// Fields are arranged to eliminate padding.
 #[derive(Debug, Copy, Clone)]
 pub(super) struct FogParams {
-    fog_color: [f32; 4],
-    fog_center: [f32; 3],
+    fog_color: [f32; 3],
     fog_start: f32,
+    fog_center: [f32; 3],
     fog_end: f32,
 }
 
@@ -12,7 +15,7 @@ implement_uniform_block!(FogParams, fog_color, fog_center, fog_start, fog_end);
 
 impl From<&GridViewRender3D<'_>> for FogParams {
     fn from(gvr3d: &GridViewRender3D<'_>) -> Self {
-        let fog_color = crate::colors::BACKGROUND_3D;
+        let fog_color = crate::colors::BACKGROUND_3D.into_raw();
 
         let fog_center = gvr3d
             .xform